Add files via upload
This is the refactoring and the testing data for the Large Class Bad Smell. Note that the testing and refactoring is done in
one step as almost all of the testing was also done during the previous iteration. Some of the testing % are lower than usual as the import statements were not tested